Area selection for processing
Restructured Fragment shader to enable image processing only for a certain area in the canvas and not for the full one. Currently coordinated have to be set manually (standard = full canvas). In future work, extend to enable the user to click / drag on canvas for area selection (e.g. by use of boxcraft.js)